[skip ci] First commit for the sphinx docs update after utils and planets namespaces have been introduced
