+++ doc/sphinx/documentation/core.rst
@@ -12,10 +12,6 @@ Name                                            Type            Description
 :class:`PyKEP.epoch`                            class           represents an epoch (i.e. a fixed point in time) 
 :func:`PyKEP.epoch_from_string`                 function        helper function to construct an epoch from a string containing a date in the format YYYY-MM-DD HH:MM:SS
 :func:`PyKEP.epoch_from_iso_string`             function        helper function to construct an epoch from a string containing a date in the ISO format YYYYMMDDTHHMMSS
-:class:`PyKEP.planet`                           class           a generic planet, ephemerides are keplerian
-:class:`PyKEP.planet_ss`                        class           a solar system planet, ephemerides are the `JPL low-precision <http://ssd.jpl.nasa.gov/txt/aprx_pos_planets.pdf>`_
-:class:`PyKEP.planet_mpcorb`                    class           represents an asteroid from the `MPCORB database <http://www.minorplanetcenter.org/iau/MPCORB.html>`_ 
-:class:`PyKEP.planet_tle`                       class           represents a satellite defined by its `TLEs <http://www.celestrak.com/NORAD/elements/>`_, ephemerides will be computed using an SGP4 propagator
 :class:`PyKEP.lambert_problem`                  class           solves the multirevolution lambert problem
 :func:`PyKEP.propagate_lagrangian`              function        propagates pure keplerian motion using Lagrange coefficients and universal variables
 :func:`PyKEP.propagate_taylor`                  function        propagates keplerian motion disturbed by a constant inertial thrust using Taylor integration method
@@ -24,9 +20,6 @@ Name                                            Type            Description
 :func:`PyKEP.fb_prop`                           function        propoagates forward a fly-by hyperbola returning the new inetrial velocity of a spacecraft after the planetary encounter
 :func:`PyKEP.barker`                            function        computes the (parabolic) time-of-flight from the Barker equation
 :func:`PyKEP.ic2par`                            function        Transforms r and v into the osculating orbital elements
-:func:`PyKEP.read_satcat`                       function        Reads the `SATCAT database <http://celestrak.com/satcat/search.asp>`_  into a dictionary 
-:func:`PyKEP.read_tle`                          function        Reads a TLE file. These can be downloaded from `Celestrack <http://www.celestrak.com/NORAD/elements/>`_ or, if you have an account, `Space Track <https://www.space-track.org/>`_
-
 =========================================       =========       ================================================
 
 
@@ -134,11 +127,3 @@ Detailed Documentation
 ------------
 
 .. autofunction:: PyKEP.ic2par(*args)
-
-------------
-
-.. autofunction:: PyKEP.read_satcat(*args)
-
-------------
-
-.. autofunction:: PyKEP.read_tle(*args)
