+++ doc/sphinx/documentation/utils.rst
@@ -1,0 +1,26 @@
+.. _utils:
+
+==================
+The utility module
+==================
+
+==================================================       =========       ================================================
+Name                                                     Type            Description
+==================================================       =========       ================================================
+:func:`PyKEP.utils.read_satcat`                          function        Reads the `SATCAT database <http://celestrak.com/satcat/search.asp>`_  into a dictionary 
+:func:`PyKEP.utils.read_tle`                             function        Reads a TLE file. These can be downloaded from `Celestrack <http://www.celestrak.com/NORAD/elements/>`_ or, if you have an account, `Space Track <https://www.space-track.org/>`_
+:func:`PyKEP.utils.load_spice_kernel`                    function        Loads in memory a kernel from the JPL SPICE toolbox (requires BUILD_SPICE option active when building from cmake)
+==================================================       =========       ================================================
+
+Detailed Documentation
+======================
+
+.. autofunction:: PyKEP.utils.read_satcat(*args)
+
+------------
+
+.. autofunction:: PyKEP.utils.read_tle(*args)
+
+------------
+
+.. autofunction:: PyKEP.utils.load_spice_kernel(*args)
